+++ frontend/src/lib/constants.ts
@@ -1,1 +1,14 @@
+'use client';
+
 export const baseUrl = process.env.NEXT_PUBLIC_API_URL;
+
+export const locales = {
+    en: "English",
+    uk: "Українська",
+}
+
+export const publicNav = {
+    home: "/",
+    wiki: "/wiki",
+    auth: "/dashboard",
+}